Revert "move global counters to a dedicated package (from badges / progress)"
This reverts commit a76377e8935456ede181ff11cd7f5cd5b37ac9ed.
