+++ achievements/internal/storages/badges/total_badges_source.go
@@ -1,11 +1,10 @@
 // SPDX-License-Identifier: BUSL-1.1
 
-package global
+package badges
 
 import (
 	"context"
 	"encoding/json"
-	"github.com/ice-blockchain/santa/achievements/internal/storages/badges"
 	messagebroker "github.com/ice-blockchain/wintr/connectors/message_broker"
 	"math"
 
@@ -13,7 +12,7 @@ import (
 	"github.com/pkg/errors"
 )
 
-func NewTotalBadgesProcessor(db tarantool.Connector) messagebroker.Processor {
+func NewBadgeProcessor(db tarantool.Connector) messagebroker.Processor {
 	return &totalBadgesSource{db: db}
 }
 
@@ -21,14 +20,14 @@ func (b *totalBadgesSource) Process(ctx context.Context, message *messagebroker.
 	if ctx.Err() != nil {
 		return errors.Wrap(ctx.Err(), "context failed")
 	}
-	achievedBadge := new(badges.AchievedBadgeMessage)
+	achievedBadge := new(AchievedBadgeMessage)
 	if err := json.Unmarshal(message.Value, achievedBadge); err != nil {
 		return errors.Wrapf(err, "totalBadgesSource: cannot unmarshall %v into %#v", string(message.Value), achievedBadge)
 	}
 	return errors.Wrapf(b.updateTotalBadgesCount(achievedBadge.Name, 1), "totalBadgesSource: failed to update total badge count for badge:%v", achievedBadge.Name)
 }
 
-func (b *totalBadgesSource) updateTotalBadgesCount(badgeName string, diff int64) error {
+func (b *totalBadgesSource) updateTotalBadgesCount(badgeName BadgeName, diff int64) error {
 	key := "TOTAL_BADGES_" + badgeName
 	op := "+"
 	if math.Signbit(float64(diff)) {
