+++ achievements/internal/storages/progress/users_source.go
@@ -47,6 +47,9 @@ func (u *userSource) handleUserDeletion(user *users.UserSnapshot) error {
 	if err := u.r.DeleteUserProgress(user.Before.ID); err != nil {
 		return errors.Wrapf(err, "failed to deleteUserAchievements")
 	}
+	if err := u.r.UpdateTotalUsersCount(-1); err != nil {
+		return errors.Wrapf(err, "failed to update total_users counter")
+	}
 
 	if user.Before.ReferredBy != "" {
 		if err := u.r.UpdateT1ReferralsCount(user.Before.ReferredBy, -1); err != nil {
@@ -60,7 +63,10 @@ func (u *userSource) handleUserDeletion(user *users.UserSnapshot) error {
 func (u *userSource) handleUserCreation(user *users.UserSnapshot) error {
 	_, err := u.r.GetUserProgress(user.ID)
 	if errors.Is(err, storage.ErrNotFound) {
-		// User's achievements record does not exists, so it is a new user - insert
+		// User's achievements record does not exists, so it is a new user - increment counter.
+		if err = u.r.UpdateTotalUsersCount(1); err != nil {
+			return errors.Wrapf(err, "failed to update total_users counter")
+		}
 		if err = u.r.InsertUserProgress(user.ID); err != nil {
 			return errors.Wrapf(err, "failed to insert user achievements record")
 		}
